+++ src/components/base/b-router/modules/const.ts
@@ -11,8 +11,7 @@
  * For instance, navigating to an external resource or using anchor-based navigation within a page.
  */
 export const urlsToIgnore = [
+	/^$/,
 	/^#/,
-	/^javascript:/,
-	/^mailto:/,
-	/^tel:/
+	/^.+?:/
 ];