+++ src/components/base/b-router/test/helpers.ts
@@ -8,30 +8,35 @@
 
 import { Component, Utils } from 'tests/helpers';
 
+import type { StaticRoutes } from 'core/router';
 import type { EngineName, InitRouter } from 'components/base/b-router/test/interface';
-
 import type { Router } from 'components/base/b-router/interface';
 
 /**
  * Returns a function to initialize the router on the page with the specified engine
+ *
  * @param engineName
+ * @param [routes]
+ * @param [props]
  */
-export function createInitRouter(engineName: EngineName): InitRouter {
+export function createInitRouter(engineName: EngineName, routes?: StaticRoutes, props?: Dictionary): InitRouter {
 	return async (page, initOptions = {}) => {
-
 		if (initOptions.initialRoute === undefined && engineName === 'in-memory') {
 			initOptions.initialRoute = 'main';
 		}
 
-		let engine: (() => Router) | undefined;
+		let
+			engine: (() => Router) | undefined;
 
 		switch (engineName) {
 			case 'history':
 				engine = Utils.evalInBrowser(() => globalThis.importModule('./src/core/router/engines/browser-history.ts').default);
 				break;
+
 			case 'in-memory':
 				engine = Utils.evalInBrowser(() => globalThis.importModule('./src/core/router/engines/in-memory.ts').default);
 				break;
+
 			default:
 				// Do nothing
 		}
@@ -41,6 +46,7 @@ export function createInitRouter(engineName: EngineName): InitRouter {
 
 			engine,
 			initialRoute: initOptions.initialRoute ?? undefined,
+			...props,
 
 			routes: {
 				main: {
@@ -134,7 +140,9 @@ export function createInitRouter(engineName: EngineName): InitRouter {
 				notFound: {
 					default: true,
 					content: '404'
-				}
+				},
+
+				...routes
 			}
 		});